Add files via upload
MistSkeleton.js contains the backend functions of Mist Tracker. At this time it is targeting Apps Scripts as a proof of concept.
MistUpdateHeaders.csv is a list of headers for various update csv types. Each line is intended to be a header for a different csv at this time.
story.rtf is a sample story used my Mist Tracker to help develop document parsing functionalities.
Viewport.html is the front end for Google Sheets.
Waterfall is a volume that contains a sample English dictionary and another copy of story.rtf.

To test in current state. Unpack Waterfall folder in a folder named "Mist Tracker", mistSkeleton.js should be placed in Waterfall folder in a new Apps Scripts file, along with viewport.html.